api/articles/id (DELETE) updated

+++ article/views.py
@@ -11,9 +11,17 @@ from .serializers import ArticleSerializer
 
 
 @csrf_exempt
-@api_view(['GET', 'POST'])
-def manage_articles(request):
+@api_view(['GET', 'POST', 'DELETE'])
+def manage_articles(request, article_id=None):
     if request.method == 'GET':
+        if article_id:
+            try:
+                articles = Article.objects.get(id=article_id)
+                serializer = ArticleSerializer(articles)
+                return Response(serializer.data)
+            except Article.DoesNotExist:
+                return Response({'error': '找不到指定的文章'}, status=status.HTTP_404_NOT_FOUND)
+
         articles = Article.objects.all()
         serializer = ArticleSerializer(articles, many=True)
         return Response(serializer.data)
@@ -26,14 +34,13 @@ def manage_articles(request):
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+    elif request.method == 'DELETE':
+        if article_id:
+            try:
+                article = Article.objects.get(id=article_id)
+                article.delete()
+                return Response({'message': '文章已成功刪除'})
+            except Article.DoesNotExist:
+                return Response({'error': '找不到指定的文章'}, status=status.HTTP_404_NOT_FOUND)
+
     return JsonResponse({'error': '請求方法不支援'})
-
-
-@api_view(['DELETE'])
-def delete_article(request, article_id):
-    try:
-        article = Article.objects.get(id=article_id)
-        article.delete()
-        return Response({'message': '文章已成功刪除'})
-    except Article.DoesNotExist:
-        return Response({'error': '找不到指定的文章'}, status=status.HTTP_404_NOT_FOUND)
